backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
SQLAlchemy-Utils==0.41.1
python-dateutil==2.9.0
beautifulsoup4==4.12.3
